Simplify package structure
The pkg directory has been removed and all its contents have been moved
to the root of the project.

The counter and report packages have been merged into a single gobble
package. This simplifies the project structure and makes the code easier
to maintain.

The main application has been updated to use the new gobble package.