Move print logic to stats package
The report package has been removed and its logic has been moved to the
stats package. The Stats struct now has a Print method, and there is a
new PrintHeader function.

This change improves the design by co-locating the data with the
logic that prints it. It also simplifies the overall structure of the
application by removing an unnecessary package.